+++ app/src/main/java/com/battlelancer/seriesguide/backend/CloudSetupFragment.kt
@@ -17,7 +17,6 @@ import com.battlelancer.seriesguide.databinding.FragmentCloudSetupBinding
 import com.battlelancer.seriesguide.sync.SgSyncAdapter
 import com.battlelancer.seriesguide.sync.SyncProgress
 import com.battlelancer.seriesguide.traktapi.ConnectTraktActivity
-import com.battlelancer.seriesguide.traktapi.TraktCredentials
 import com.battlelancer.seriesguide.util.Errors
 import com.battlelancer.seriesguide.util.Utils
 import com.battlelancer.seriesguide.util.safeShow
@@ -269,14 +268,6 @@ class CloudSetupFragment : Fragment() {
     }
 
     private fun updateViews() {
-        // warn about changes in behavior with trakt
-        binding?.textViewCloudWarnings?.visibility =
-            if (TraktCredentials.get(activity).hasCredentials()) {
-                View.VISIBLE
-            } else {
-                View.GONE
-            }
-
         // hexagon enabled and account looks fine?
         if (HexagonSettings.isEnabled(context)
             && !HexagonSettings.shouldValidateAccount(context)) {