+++ app/src/main/java/com/battlelancer/seriesguide/ui/preferences/SgPreferencesFragment.kt
@@ -35,7 +35,7 @@ import com.battlelancer.seriesguide.streaming.StreamingSearch
 import com.battlelancer.seriesguide.streaming.StreamingSearchConfigureDialog
 import com.battlelancer.seriesguide.sync.SgSyncAdapter
 import com.battlelancer.seriesguide.ui.SeriesGuidePreferences
-import com.battlelancer.seriesguide.ui.dialogs.LanguageChoiceDialogFragment
+import com.battlelancer.seriesguide.ui.dialogs.ShowL10nDialogFragment
 import com.battlelancer.seriesguide.ui.dialogs.NotificationSelectionDialogFragment
 import com.battlelancer.seriesguide.ui.dialogs.NotificationThresholdDialogFragment
 import com.battlelancer.seriesguide.ui.dialogs.TimeOffsetDialogFragment
@@ -214,9 +214,8 @@ class SgPreferencesFragment : PreferenceFragmentCompat(),
         findPreference<Preference>(DisplaySettings.KEY_LANGUAGE_FALLBACK)!!.also {
             updateFallbackLanguageSummary(it)
             it.setOnPreferenceClickListener {
-                LanguageChoiceDialogFragment.show(
+                ShowL10nDialogFragment.show(
                     parentFragmentManager,
-                    R.array.languageCodesShows,
                     DisplaySettings.getShowsLanguageFallback(context),
                     TAG_LANGUAGE_FALLBACK
                 )
@@ -459,7 +458,7 @@ class SgPreferencesFragment : PreferenceFragmentCompat(),
     }
 
     @Subscribe(threadMode = ThreadMode.MAIN)
-    fun onEvent(event: LanguageChoiceDialogFragment.LanguageChangedEvent) {
+    fun onEvent(event: ShowL10nDialogFragment.LanguageChangedEvent) {
         if (event.tag == TAG_LANGUAGE_FALLBACK) {
             PreferenceManager.getDefaultSharedPreferences(context).edit {
                 putString(DisplaySettings.KEY_LANGUAGE_FALLBACK, event.selectedLanguageCode)